achat express quantité =1 ; création de l'entrée dans la BDD Order. mise à jour quantité en stock KO

+++ app/controllers/checkout_controller.rb
@@ -23,7 +23,8 @@ class CheckoutController < ApplicationController
       ],
       metadata: {
         joint_table_cart_book_ids: joint_table_cart_book_ids.join(','),
-        user_id: @user_id
+        user_id: @user_id,
+        checkout_type: 'cart_checkout',
        
       },
       mode: 'payment',
@@ -65,6 +66,8 @@ class CheckoutController < ApplicationController
 
       metadata: {
         user_id: @user_id,
+        checkout_type: 'express_checkout',
+        book_id: @book.id,
       },
 
       mode: 'payment',
@@ -79,27 +82,43 @@ class CheckoutController < ApplicationController
   end
 
   def success
-
     @session = Stripe::Checkout::Session.retrieve(params[:session_id])
     @payment_intent = Stripe::PaymentIntent.retrieve(@session.payment_intent)
-    joint_table_cart_book_ids = @session.metadata.joint_table_cart_book_ids.split(',').map(&:to_i)
-   
+    checkout_type = @session.metadata['checkout_type']
 
+    if checkout_type == 'cart_checkout' #cas du paiement via le panier
+      joint_table_cart_book_ids = @session.metadata.joint_table_cart_book_ids.split(',').map(&:to_i)
+  
     #Création d'une nouvelle instance dans la BDD Order (voir dans le model Order)
-    Order.create_order_with_books(current_user, joint_table_cart_book_ids)
+       Order.create_order_with_books(current_user, joint_table_cart_book_ids)
 
     #Gérer la màj des quantités après un paiement stripe:
-    joint_table_cart_book_ids.each do |joint_table_cart_book_id| #conversion des identifiants des JointTableCartBook, on les diviseàchaque virgule, on converti en entiers
-      joint_table_cart_book = JointTableCartBook.find(joint_table_cart_book_id) #on chercheun id specifique
-  
-      # Ajouter une logique pour mettre à jour la quantité en stock du livre
-      book = joint_table_cart_book.book 
-      new_stock_quantity = book.quantity - joint_table_cart_book.quantity #new stock = quantité actuelle du livre - qté commandée
-      book.update(quantity: new_stock_quantity) #mise à jour de la qté en stock
-    end
+       joint_table_cart_book_ids.each do |joint_table_cart_book_id| #conversion des identifiants des JointTableCartBook, on les diviseàchaque virgule, on converti en entiers
+        joint_table_cart_book = JointTableCartBook.find(joint_table_cart_book_id) #on chercheun id specifique
+      # Mettre à jour la quantité en stock du livre
+        book = joint_table_cart_book.book 
+        new_stock_quantity = book.quantity - joint_table_cart_book.quantity #new stock = quantité actuelle du livre - qté commandée
+        book.update(quantity: new_stock_quantity) #mise à jour de la qté en stock
+      end
 
     # Effacer les articles du panier après la commande
       current_user.cart.joint_table_cart_books.destroy_all
+
+
+    else # gestion pour l'achat express (pas de panier)
+      book_id = params[:book_id]
+      user_id = @session.metadata.user_id
+      total_price = @session.amount_total / 100.0 
+
+       # Récupérer le livre associé à l'achat express
+      book = Book.find_by(id: book_id)
+
+      #Création d'une nouvelle instance dans la BDD Order (voir dans le model Order)
+      Order.create_order_for_express_purchase(current_user, book, total_price)
+
+      #Gérer la màj des quantités après un paiement stripe:
+      #book.update_stock_quantity(1)
+    end
   end
 
   def cancel
